Simplified undo stack.
A single entry in the undo stack is now always a single undo_data_t, not a vector. If multiple data need to be saved, it should all be part of a single undo_data_t. Since having a single item was the vast majority of use case, paying the allocation penalty of a vector was not worth it.