Prevent swapping or assigning to a sliced reference.
This is to prevent swapping between a ref_t and valid_ref_t, making the valid reference invalid.

This is also to prevent swapping between a ref_t or valid_ref_t and edit_ref_t, making the edit refernce refers to an object that has not been added to a transaction.