Split any_op::op::call into op::find and op::call.
This allows caller to do the find themselves and try multiple finds.

This is used in the ref stream op to use the non-ref stream op if the ref-stream op does not exists. This avoids having to register a ref-stream op for types that already have a stream-op that does adequately the job.