Simplifying object design.
The goal is to rely less on any_op. The problem with any_op is that it does not work well with super-types, the op-searching mechanism cannot easily walk the type tree of multiple arguments without being extremely complicated, unreadable template code.

The goal is also to promote using object types as -is.

- Move type info to utility library.
- Removed the constant base class of object to simplify.
- Made value favor core types.
- Made value keep all references as references to object.
- Made commands be objects.